Change how the pressure and position logic inside the device work and move the InputInjector into the touch processor

+++ TouchPanels/Devices/ITouchDevice.cs
@@ -9,7 +9,7 @@ namespace TouchPanels.Devices
 {
     public interface ITouchDevice
     {
-        int Pressure { get; }
+        double Pressure { get; }
         Point TouchPosition { get; }
         void ReadTouchpoints();
         Point RawTouchPosition { get; }
